Add logout functionality back for #79
Clear cookies on path=/ for #62.
