+++ client/src/App.tsx
@@ -3,7 +3,7 @@ import React, { CSSProperties } from 'react'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 import { colors, classes } from './styles'
 // import { useWarnResize } from './warnResize'
-import { useRefreshToken } from './fetchWrapper'
+import { useRefreshToken } from './auth'
 import { PlaylistGrid } from './PlaylistGrid'
 import { PlaylistEditor } from './Playlist-Editor/PlaylistEditor'
 import { Header } from './Header'
